Extract icons to separate files
