+++ src/assets/icons/chevronRight.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+export const chevronRight = (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="50"
+    height="50"
+    viewBox="0 0 512 512"
+  >
+    <polyline
+      points="184 112 328 256 184 400"
+      style={{
+        fill: 'none',
+        strokeLinecap: 'round',
+        strokeLinejoin: 'round',
+        strokeWidth: '48px',
+      }}
+    />
+  </svg>
+);
