+++ src/assets/icons/index.ts
@@ -1,0 +1,7 @@
+export * as tools from './tools/index';
+export { chevronRight } from './chevronRight';
+export { codepen } from './codepen';
+export { email } from './email';
+export { linkedIn } from './linkedin';
+export { twitter } from './twitter';
+export { github } from './github';
