+++ src/assets/scss/_variables.scss
@@ -81,4 +81,4 @@ $font-size-md: 16px;
 $font-size-lg: 18px;
 $font-size-xl: 20px;
 $font-size-xxl: 22px;
-$font-size-heading: 40px;
+$font-size-heading: 25px;
