+++ src/components/sections/about/index.tsx
@@ -1,46 +1,29 @@
 import React from 'react';
 import { StaticImage } from 'gatsby-plugin-image';
 
-import {
-  docker,
-  typescript,
-  react,
-  vuejs,
-  scss,
-  k8s,
-  postgres,
-  nodejs,
-  graphql,
-  html5,
-  mongodb,
-  javascript,
-  firebase,
-  tailwind,
-  git,
-  jest,
-} from '../../../assets/icons/icons';
+import { tools } from '../../../assets/icons';
 import './about.scss';
 import { AppLink } from '../../AppLink';
 
 export interface AboutProps {}
 
 const skills = [
-  { name: 'HTML', icon: html5 },
-  { name: 'CSS/SCSS', icon: scss },
-  { name: 'TailwindCSS', icon: tailwind },
-  { name: 'Javascript', icon: javascript },
-  { name: 'Typescript', icon: typescript },
-  { name: 'VueJS', icon: vuejs },
-  { name: 'React/React Native', icon: react },
-  { name: 'NodeJS', icon: nodejs },
-  { name: 'Firebase', icon: firebase },
-  { name: 'GraphQL', icon: graphql },
-  { name: 'Postgres', icon: postgres },
-  { name: 'MongoDB', icon: mongodb },
-  { name: 'Docker', icon: docker },
-  { name: 'Kurbernetes', icon: k8s },
-  { name: 'Git', icon: git },
-  { name: 'Software Testing', icon: jest },
+  { name: 'HTML', icon: tools.html5 },
+  { name: 'CSS/SCSS', icon: tools.scss },
+  { name: 'TailwindCSS', icon: tools.tailwind },
+  { name: 'Javascript', icon: tools.javascript },
+  { name: 'Typescript', icon: tools.typescript },
+  { name: 'VueJS', icon: tools.vuejs },
+  { name: 'React/React Native', icon: tools.react },
+  { name: 'NodeJS', icon: tools.nodejs },
+  { name: 'Firebase', icon: tools.firebase },
+  { name: 'GraphQL', icon: tools.graphql },
+  { name: 'Postgres', icon: tools.postgres },
+  { name: 'MongoDB', icon: tools.mongodb },
+  { name: 'Docker', icon: tools.docker },
+  { name: 'Kurbernetes', icon: tools.k8s },
+  { name: 'Git', icon: tools.git },
+  { name: 'Software Testing', icon: tools.jest },
 ];
 
 export const About: React.FC<AboutProps> = () => {
