+++ src/layouts/main/layout.scss
@@ -11,7 +11,7 @@ body {
 .site-wrapper {
   max-width: 1500px;
   margin: auto;
-  padding: 0 5rem;
+  padding: 0 10rem;
 }
 
 @media screen and (max-width: $screen-lg-max) {
